LayoutTests/http/tests/security/contentSecurityPolicy/connect-src-websocket-blocked-expected.txt

CONSOLE MESSAGE: Refused to connect to 'ws://localhost:8880/websocket/tests/hybi/echo' because it violates the following Content Security Policy directive: "connect-src ws://127.0.0.1:8880".

Pass
